Add reservation permissions

+++ module/ShopBundle/Controller/ShopController.php
@@ -18,7 +18,8 @@
 
 namespace ShopBundle\Controller;
 
-use Zend\View\Model\ViewModel;
+use DateTime,
+    Zend\View\Model\ViewModel;
 
 /**
  * ShopController
@@ -29,8 +30,7 @@ class ShopController extends \CommonBundle\Component\Controller\ActionController
 {
     public function indexAction()
     {
-        //TODO
-        $canReserve = true;
+        $canReserve = $this->canReserve();
 
         return new ViewModel(
             array(
@@ -41,38 +41,50 @@ class ShopController extends \CommonBundle\Component\Controller\ActionController
 
     public function reserveAction()
     {
-        $reserveForm = $this->getForm('shop_shop_reserve');
-        $canReserve = true;
+        $canReserve = $this->canReserve();
+        if (!$canReserve) {
+            $this->flashMessenger()->error(
+                'Error',
+                'You are not allowed to make reservations!'
+            );
 
+            return new ViewModel();
+        }
+
+        $reserveForm = $this->getForm('shop_shop_reserve');
         if ($this->getRequest()->isPost()) {
             $formData = $this->getRequest()->getPost();
             $reserveForm->setData($formData);
 
             if ($reserveForm->isValid()) {
                 $reservation = $reserveForm->hydrateObject();
-                $reservation->setPerson($this->getPersonEntity());
-                $this->getEntityManager()->persist($reservation);
-                $this->getEntityManager()->flush();
+                if ($reservation->getSalesSession()->getStartDate() <= new DateTime()) {
+                    $this->flashMessenger()->error(
+                        'Error',
+                        'You can only make reservations for sales sessions that have not started yet.'
+                    );
+                } else {
+                    $reservation->setPerson($this->getPersonEntity());
+                    $this->getEntityManager()->persist($reservation);
+                    $this->getEntityManager()->flush();
 
-                $this->flashMessenger()->success(
-                    'Success',
-                    'The reservation was successfully made!'
-                );
-
-                $this->redirect()->toRoute(
-                    'shop',
-                    array(
-                        'action' => 'reserve',
-                    )
-                );
-
-                return new ViewModel();
+                    $this->flashMessenger()->success(
+                        'Success',
+                        'The reservation was successfully made!'
+                    );
+                }
             } else {
                 $this->flashMessenger()->error(
                     'Error',
                     'An error occurred while processing your reservation!'
                 );
             }
+            $this->redirect()->toRoute(
+                'shop',
+                array(
+                    'action' => 'reservations',
+                )
+            );
         }
 
         return new ViewModel(
@@ -85,8 +97,7 @@ class ShopController extends \CommonBundle\Component\Controller\ActionController
 
     public function reservationsAction()
     {
-        //TODO
-        $canReserve = true;
+        $canReserve = $this->canReserve();
 
         $reservations = $this->getEntityManager()
             ->getRepository('ShopBundle\Entity\Reservation')
@@ -103,12 +114,19 @@ class ShopController extends \CommonBundle\Component\Controller\ActionController
     public function deleteReservationAction()
     {
         if ($reservation = $this->getEntityById('ShopBundle\Entity\Reservation')) {
-            $this->getEntityManager()->remove($reservation);
-            $this->getEntityManager()->flush();
+            $canBeDeleted = true;
+            $canBeDeleted = $canBeDeleted && $reservation->getPerson()->getId() == $this->getPersonEntity()->getId();
+            $canBeDeleted = $canBeDeleted && $reservation->getSalesSession()->getStartDate() > new DateTime();
+            if (!$canBeDeleted) {
+                $this->flashMessenger()->error('Error', 'You don\'t have permission to cancel this reservation.');
+            } else {
+                $this->getEntityManager()->remove($reservation);
+                $this->getEntityManager()->flush();
 
-            $this->flashMessenger()->success("Success", "Your reservation was successfully cancelled");
+                $this->flashMessenger()->success('Success', 'Your reservation was successfully cancelled');
+            }
         } else {
-            $this->flashMessenger()->error("Error", "An error occurred while trying to cancel your reservation");
+            $this->flashMessenger()->error('Error', 'An error occurred while trying to cancel your reservation');
         }
 
         $this->redirect()->toRoute(
@@ -131,5 +149,99 @@ class ShopController extends \CommonBundle\Component\Controller\ActionController
         }
 
         return $this->getAuthentication()->getPersonObject();
+    }
+
+    /**
+	 * @return bool
+	 */
+    private function canReserve()
+    {
+        if (!$this->getAuthentication()->isAuthenticated()) {
+            return false;
+        }
+
+        //reservation permissions
+        $reservationPermission = $this->getEntityManager()
+            ->getRepository('ShopBundle\Entity\ReservationPermission')
+            ->find($this->getPersonEntity());
+        if ($reservationPermission) {
+            return $reservationPermission->getReservationsAllowed();
+        }
+
+        $configRepository = $this->getEntityManager()
+            ->getRepository('CommonBundle\Entity\General\Config');
+
+        //default permission
+        if ($configRepository->getConfigValue('shop.reservation_default_permission')) {
+            return true;
+        }
+
+        //organization role
+        if ($configRepository->getConfigValue('shop.reservation_organisation_status_permission_enabled')) {
+            $status = $this->getPersonEntity()->getOrganizationStatus($this->getCurrentAcademicYear());
+            if ($status->getStatus() == $configRepository->getConfigValue('shop.reservation_organisation_status_permission_status')) {
+                return true;
+            }
+        }
+
+        //total shifts
+        if ($configRepository->getConfigValue('shop.reservation_shifts_general_enabled')) {
+            if ($this->getTotalShiftCount() >= $configRepository->getConfigValue('shop.reservation_shifts_general_number')) {
+                return true;
+            }
+        }
+
+        //shifts for unit
+        if ($configRepository->getConfigValue('shop.reservation_shifts_permission_enabled')) {
+            if ($this->getUnitShiftCount($configRepository->getConfigValue('shop.reservation_shifts_unit_id')) >= $configRepository->getConfigValue('shop.reservation_shifts_number')) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    /**
+	 * @return int
+	 */
+    private function getTotalShiftCount()
+    {
+        $shiftCount = 0;
+        $now = new DateTime();
+        $shifts = $this->getEntityManager()
+            ->getRepository('ShiftBundle\Entity\Shift')
+            ->findAllByPersonAsVolunteer($this->getPersonEntity(), $this->getCurrentAcademicYear());
+        foreach ($shifts as $shift) {
+            if ($shift->getStartDate() > $now) {
+                continue;
+            }
+            $shiftCount++;
+        }
+
+        return $shiftCount;
+    }
+
+    /**
+	 * @param $unitId
+	 * @return int
+	 */
+    private function getUnitShiftCount($unitId)
+    {
+        $shiftCount = 0;
+        $now = new DateTime();
+        $shifts = $this->getEntityManager()
+            ->getRepository('ShiftBundle\Entity\Shift')
+            ->findAllByPersonAsVolunteer($this->getPersonEntity(), $this->getCurrentAcademicYear());
+        foreach ($shifts as $shift) {
+            if ($shift->getStartDate() > $now) {
+                continue;
+            }
+            if ($shift->getUnit()->getId() != $unitId) {
+                continue;
+            }
+            $shiftCount++;
+        }
+
+        return $shiftCount;
     }
 }
